+++ core/src/routes/(modules)/options/div/flexbox-grid-basis.svelte
@@ -43,7 +43,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';