+++ core/src/routes/(modules)/options/div/flexbox-grid-flex-shrink.svelte
@@ -49,7 +49,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';