+++ core/src/routes/(modules)/options/div/flexbox-grid-flex-wrap.svelte
@@ -50,7 +50,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';