+++ core/src/routes/(modules)/options/div/flexbox-grid-flex.svelte
@@ -51,7 +51,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';