+++ core/src/routes/(modules)/options/div/flexbox-grid-justify-content.svelte
@@ -55,7 +55,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';