+++ core/src/routes/(modules)/options/div/flexbox-grid-justify-self.svelte
@@ -53,7 +53,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';