+++ core/src/routes/(modules)/options/div/flexbox-grid-place-items.svelte
@@ -52,7 +52,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';