+++ core/src/routes/(modules)/options/div/flexbox-grid-rows-span.svelte
@@ -63,7 +63,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';