+++ core/src/routes/(modules)/options/div/flexbox-grid-rows-start-end.svelte
@@ -78,7 +78,7 @@
     let propertyList = [];
 
     propertyOptions.forEach(elm => {
-        propertyList.push(elm.value);
+        if(elm.value != "") propertyList.push(elm.value);
     });
 
     import { createEventDispatcher } from 'svelte';